code/requirements.txt
dotenv
pandas
tqdm

# Download Data Tool
yfinance
requests
bs4

# Data Load Tool with S3 + Qdrant
dlt[qdrant,s3]>=1.4.0  
dlt[duckdb]

# AWS SDK for Python
boto3

# vector database
qdrant-client[fastembed]
fastembed

# database
duckdb

# llm
google-genai

# UI
streamlit
